temporary but _definitely NOT_ ideal fix, music now works.
i did not know this previously, but the way doom plays a music track
is by first extracting the MIDI file from the WAD to a temporary
file (doom.mid) and then playing it from there.

for a wii, this will cause unnecessary wear-and-tear on the players SD
card, so this behaviour should be changed (probably will end up
extracting all MIDI tracks from the WAD somewhere at startup if we
detect that it has not already been done so).